Remove some redundancy & add Quote to bridge
